added classes arc_lifting_trihedral_pair, vector_activity, vector_activity_description

+++ src/lib/layer3_group_actions/induced_actions/induced_actions.h
@@ -711,7 +711,7 @@ public:
 	algebra::basic_algebra::matrix_group *M;
 	algebra::field_theory::finite_field *F;
 	int low_level_point_size;
-	int degree;
+	long int degree;
 
 	// wedge product
 	int dimension; // = HPD->nb_monomials
